+++ src/main/java/its_meow/betteranimalsplus/proxy/CommonProxy.java
@@ -6,6 +6,7 @@ import its_meow.betteranimalsplus.common.config.BetterAnimalsPlusConfig;
 import its_meow.betteranimalsplus.common.world.gen.TrilliumGenerator;
 import its_meow.betteranimalsplus.init.BlockRegistry;
 import its_meow.betteranimalsplus.init.CraftingRegistry;
+import its_meow.betteranimalsplus.init.MobRegistry;
 import net.minecraftforge.common.config.Configuration;
 import net.minecraftforge.fml.common.event.FMLInitializationEvent;
 import net.minecraftforge.fml.common.event.FMLPostInitializationEvent;
@@ -17,6 +18,7 @@ public class CommonProxy {
 	public static Configuration config;
 
 	public void preInit(FMLPreInitializationEvent event) {
+		MobRegistry.fillContainers();
 		File directory = event.getModConfigurationDirectory();
         config = new Configuration(new File(directory.getPath(), "betteranimalsplus.cfg")); 
         BetterAnimalsPlusConfig.readConfig();